fileの名前変更、mysql接続とレコード検索の責務を関数単位で分割
