--- api/mysql_set/db_connect.go
@@ -1,11 +1,0 @@
-package mysql_set
-
-import (
-
-)
-
-func Mysql_Connect() *SQLHandler {
-	DBOpen()
-	DB := GetDBConn()
-	return DB
-}
