+++ api/mysql_setting/db_connect.go
@@ -1,0 +1,38 @@
+package mysql_setting
+
+import (
+	"gorm.io/gorm"
+)
+
+// DBOpen は DB connectionを張る。
+func DBOpen() {
+	dbConn = NewSQLHandler()
+}
+
+// DBClose は DB connectionを張る。
+func DBClose() {
+	sqlDB, _ := dbConn.DB.DB()
+	sqlDB.Close()
+}
+
+func Mysql_Connect() *SQLHandler {
+	DBOpen()
+	DB := GetDBConn()
+	return DB
+}
+
+// GetDBConn ...
+func GetDBConn() *SQLHandler {
+	return dbConn
+}
+
+// BeginTransaction ...
+func BeginTransaction() *gorm.DB {
+	dbConn.DB = dbConn.DB.Begin()
+	return dbConn.DB
+}
+
+// Rollback ...
+func RollBack() {
+	dbConn.DB.Rollback()
+}
